addressed issues
change test file name to match naming convention.
renamed _get_headers to be private and made the return strict.
removed main from ollama_api.py.
improved generate_prompt_response with default values, directly access the "response" key in the return.

hope this is the last pr before merging, sorry.

the password is something you get at the doctors office when you might have broken something. lowercase
